correctly apllied to landing page now it will be easir for internal components
