+++ src/components/ui/landingPage/MarketingPage.tsx
@@ -23,22 +23,22 @@ import { light } from "@mui/material/styles/createPalette";
 export default function MarketingPage() {
   const [mode, setMode] = React.useState<PaletteMode>("light");
   const [mounted, setMounted] = React.useState(false);
-  const [showCustomTheme, setShowCustomTheme] = React.useState(true);
+  // const [showCustomTheme, setShowCustomTheme] = React.useState(true);
 
   // This code only runs on the client side, to determine the system color preference
   React.useEffect(() => {
     setMounted(true);
     // Check if there is a preferred mode in localStorage
-    const savedMode = localStorage.getItem("themeMode") as PaletteMode | null;
-    if (savedMode) {
-      setMode(savedMode);
-    } else {
-      // If no preference is found, it uses system preference
-      const systemPrefersDark = window.matchMedia(
-        "(prefers-color-scheme: dark)"
-      ).matches;
-      setMode(systemPrefersDark ? "dark" : "light");
-    }
+    // const savedMode = localStorage.getItem("themeMode") as PaletteMode | null;
+    // if (savedMode) {
+    //   setMode(savedMode);
+    // } else {
+    //   // If no preference is found, it uses system preference
+    //   const systemPrefersDark = window.matchMedia(
+    //     "(prefers-color-scheme: dark)"
+    //   ).matches;
+    //   setMode(systemPrefersDark ? "dark" : "light");
+    // }
   }, []);
 
   // Prevent hydration mismatch
@@ -46,8 +46,8 @@ export default function MarketingPage() {
     return null;
   }
 
-  const MPTheme = createTheme(getMPTheme(mode));
-  const defaultTheme = createTheme({ palette: { mode } });
+  // const MPTheme = createTheme(getMPTheme(mode));
+  // const defaultTheme = createTheme({ palette: { mode } });
 
   const toggleColorMode = () => {
     const newMode = mode === "dark" ? "light" : "dark";
@@ -55,9 +55,9 @@ export default function MarketingPage() {
     localStorage.setItem("themeMode", newMode); // Save the selected mode to localStorage
   };
 
-  const toggleCustomTheme = () => {
-    setShowCustomTheme((prev) => !prev);
-  };
+  // const toggleCustomTheme = () => {
+  //   setShowCustomTheme((prev) => !prev);
+  // };
 
   // const theme = createTheme({
   //   palette: {
@@ -74,10 +74,9 @@ export default function MarketingPage() {
   // });
 
   return (
-    // <ThemeProvider theme={theme}>
-    <ThemeProvider theme={showCustomTheme ? MPTheme : defaultTheme}>
+    <>
+      {/* <ThemeProvider theme={showCustomTheme ? MPTheme : defaultTheme}> */}
       <CssBaseline enableColorScheme />
-
       <AppAppBar mode={mode} toggleColorMode={toggleColorMode} />
       <Hero />
       <div>
@@ -106,6 +105,7 @@ export default function MarketingPage() {
         <Divider /> */}
         <Footer mode={mode} />
       </div>
-    </ThemeProvider>
+      {/* </ThemeProvider> */}
+    </>
   );
 }
