--- src/theme/customTheme.ts
@@ -1,10 +1,0 @@
-"use client";
-import { createTheme } from "@mui/material/styles";
-
-const customTheme = createTheme({
-  typography: {
-    fontFamily: "var(--font-roboto)",
-  },
-});
-
-export default customTheme;
