+++ tsconfig.json
@@ -28,7 +28,11 @@
   "exclude": ["node_modules"],
   "ts-node": {
     "compilerOptions": {
-      "module": "CommonJS"
+      "module": "CommonJS",
+      "baseUrl": "src",
+      "paths": {
+        "@/*": ["*"]
+      }
     }
   },
 }
